modules/plugins system using rainTPL
You can now place raintpl template elements in tpl/plugins/
This allows adding custom HTML elements to the base template.
It is possible to inject these elements at the following places on the
default template. Where they are loaded depends on their path/filename.

 * Toolbar buttons (eg. tpl/plugins/coolbutton/coolbutton.toolbar.html)
 * Link info fields (eg. tpl/plugins/send-to-twitter.html)
 * Page footer (eg. tpl/plugins/footer-message/footer-message.footer.html)

Plugins have to be manually enabled in data/options.php. If you want to enable
the 'example1' and 'example2' plugins:

    $GLOBALS['config']['PLUGINS'] = array('example1', 'example2');

Some cleanup: Move rendering the linklist to a separate function;
Create a PageBuilder **every time** we render the page.

Credits to @pikzen for the refactoring and plugin loading code.